+++ src/anonlm/__init__.py
@@ -2,13 +2,14 @@
 
 from anonlm.api import anonymize, create_engine, deanonymize
 from anonlm.config import AnonLMConfig, ProviderPreset
-from anonlm.engine import AnonymizationEngine, AnonymizationResult
+from anonlm.engine import AnonymizationEngine, AnonymizationResult, ChunkingMetadata
 from anonlm.schema import PIIEntity, PIIResponse, PIIType
 
 __all__ = [
     "AnonLMConfig",
     "AnonymizationEngine",
     "AnonymizationResult",
+    "ChunkingMetadata",
     "PIIEntity",
     "PIIResponse",
     "PIIType",